% Auto-generated: do not edit by hand
\name{dashComponentUnload}

\alias{dashComponentUnload}

\title{DashComponentUnload component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
dashComponentUnload(id=NULL, close=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{close}{Logical. If the event has triggered.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

